Fill Layers: fix previewing and rendering
Both setDirty() and resetCache() are booby-trapped -- if invoked during
a preview stroke, they trigger a background update of the layer using
the existing (blank) configuration. As the update() resets the
projection of the layer for a *third* time, this breaks the in-progress
render of the preview.

As layers are rendered via update() or the new previewWithStroke(),
this commit fixes this by defanging the two functions above.