Add createtarball to our packaging repo
releaseme doesn't work for Krita because it's very obviously meant
for releasing a large number of repos where nobody bothers to prepare
tags. We cannot release from a branch, but must release from a tag.

The script still needs to be fixed by fetching the tagged tarball
from invent, unpacking and then adding the translations.