Fix layers docker not updating blendmode availability on image cs change
Blending modes available to the layer depend on the color space of the
parent layer (that is the color space where actual blending happens).
Therefore we should connect updateUI to the color space change of the
parent layer.

WARNING: in Krita **all** group layers have the color space of the image,
         you cannot convert them into any specific color space. That is
         a wish-bug that can be fixed.

BUG:374386