`pixelDataAtTime` Implementation Correction
An incorrect call to KisRasterKeyframeChannel::fetchFrame was causing undefined behavior
when reading from a specific frame. Instead of passing in a write-target paint device,
it was using the existing node's paint device which was causing the frame at current time
to be overwritten by essentially arbitrary frame data.

This user-error is easy to make with the original method name ('fetchFrame' doesn't sound
like a method that writes to a paint device, which made argument 2 somewhat confusing
from an outsiders perspective) and has already been addressed in our Refactor, but it's
probably worth fixing this first in an isolated commit.

Additional Note: the paint device being made to write to is using a snapshot copy to create
the temporary device before feeding into the fetchFrame method. Since we basically just
want an empty 'sheet' to write to, the contents of the frame should be mostly arbitrary before
writing.

BUG:409414