Fix updates when changing visibility of a layer
Well, there were three approaches to fix this bug, bad, very bad and
extremely bad. I chose the first one (see a comment
in NodeView::dataChanged).

The problem is that Qt uses visualRect() for both, drop indicator
rendering and updates of the index. And since we draw the visibility
sign to the left of the actual element, there is a problem arises.

Perhaps we could refactor our code and use "columns" feauture of
the tree view instead of rendering everything in one column. But I
have a feeling that I tried this approach when rewriting the docker
the last time, and it didn't work for some reason.