Animation: Clone Frames
Adds the ability to create **Clone Frames**, raster keyframes that point to the exact same piece of physical frame data inside of a KisPaintDevice. An often requested feature, clone frames allow the animator to easily reuse an existing keyframe throughout their animation and (unlike a duplicated frame) changes made to one instance are reflected in all of its clones.

- Create a clone by `holding ALT`, `Left-Click Dragging` a raster keyframe, and `Releasing Left-Click over another frame slot`.

- Added a `Clone Keyframes` context menu action to the frames view grid. This works just like copying keyframes to the clipboard, but the pasted frames will be clones!

- Added a `Make Unique` context menu action that exists when right-clicking on a selection containing one or more clone frames. This allows users to convert any clone frame into a regular duplicate.

- The `Shift Keyframes` hotkey has been moved from ALT to SHIFT.

(Designed and programmed with @eoinoneill.)