Fix full-pressure blobs when using floating dockers
On Windows tablet events may arrive asynchronously to the
mouse events (in WinTab mode). The problem is that Qt
generates Enter/Leave and FocusIn/Out events via mouse
events only. It means that TabletPress may come much before
Enter and FocusIn event and start the stroke. In such a case
we shouldn't unblock mouse events.

PS:
Ideally, we should fix Qt to generate Enter/Leave and
FocusIn/Out events based on tablet events as well, but
it is a lot of work.

BUG:417040