Fix bad rendering of text shapes
Before this commit, white text over white canvas
would be visible because of artifacts.
This commit fixes it and now white text over white canvas is invisible
in case there is no additional "stroke".

WARNING: it changes the rendering of old files!
(however one could argue that the previous rendering was wrong)

BUG:418141