Fix font options after clearing the rich text editor
Before this commit, if the rich text editor for SVG Text Tool
was cleared, the options for the text would be lost and changed
to the default ones. This commit fixes it by remembering the old options
and restoring them.

Note: it might be beneficial to consider figuring out how to set the
Format in the location just before the text instead, because that would
(partially) fix bug 411739.

BUG:411393