Fix crash in ColorSmudge paintop with animated GIH brushes
The crash happened because dabRect and actual dab were calculated for
different brush indexes in the animated brush.

Before the patch, the switch of the index happened implicitly in
a call to KisBrush::brushIndex(). Now there are no implicit switches,
instead, the user is obliged to call KisBrush::prepareForSeqNo() before
doing brush generation.

Most of the times, tracking of seqNo is handled by the dab caching
system.

BUG:425875