Fix text marked modified when Text Editor is opened
Before this commit, text would be marked modified if you opened
a document with text and just opened the Text Editor.
This was a regression from a24aa5c7 (a fix for bug 411393).
This commit fixes that behaviour.

Additionally I renamed setShape to setInitialShape because the fix
relies on the function being called only once, at the very opening
of the editor.

CCBUG:411393