Fix strike-through in converstion rich text -> svg
Before this commit, strike-through would work if the user
adds `line-decoration=line-through` in svg, but it doesn't
work if the user uses the strike-through option in "Format"
submenu, because rich text -> svg conversion didn't keep it.
This commit fixes it and if the user does it, the strike-through
appears both in canvas and in the rendered svg.

Note: the bug regarding weird painting of the strike-through
on the canvas (parts inside black letters are white) is still
present both before and after the bug fix.

BUG:426608