Bugfix: Vector/references don't get rendered
KisExternalLayer had no key function, so the typeinfo
was emitted as a weak symbol. So, it may not always be
possible to merge the typeinfo with shared libraries,
correctly. This would make dynamic_cast fail (because
it depends on RTTI in this case).

src: https://android.googlesource.com/platform/ndk/+/master/docs/user/common_problems.md#rtti_exceptions-not-working-across-library-boundaries

BUG:422312