Fix crash when undoing Rectangle Selection and doing redo after
It might happen that by the time ChangeShapeSelectionCommand
is destroyed, the parent selection doesn't exist anymore. E.g.
when all the commands have been undone and are now just being
destroyed in a batch.

In such a case we should track the availability of the selection
with a weak pointer.