Add logs to the bug report dialog
Remove the compiler label; it's not filled in. And add a text
edit field with the combined sysinfo, log and crash log files,
and place the contents on the clipboard for easy pasting.

It's a pity we cannot use the bugzilla REST api to automatically
add these files as an attachment.

From now on, we can simply ask people to report bugs using the
dialog and add this information.

CCMAIL:kimageshop@kde.org
(cherry picked from commit 31b00a47a487ced1f44619cfeecd5494a8886b31)