Fix calculating lengths in SVG files
Before this commit, the same SVG file could look different
(besides the scale) depending on the resolution set by the user.
For example if there was an SVG file with
`width="100px" height="100px"`, the resulting file would always have
100x100 pixels canvas, but depending on the resolution set by the user
the lengths inside specified for example with 'em' or 'in' would
differ.

(See the example file https://www.w3.org/TR/SVG2/coords.html#Units
and test with opening the file in a new canvas with 10 and 300 dpi;
you'll notice that in one case, some rectangles are really small
on the canvas, while in the other case the same rectangles are
really huge. That makes the same SVG look very differently
depending on DPI).

After this commit, the behaviour is as follows:

1. If the SVG is imported into a new image,
    a.  if the width and height are defined in pixels "px", then the
      resulting canvas will be exactly in that size (which is consistent
      with behaviour before this commit), and the elements will have
      the same (proper) sizes no matter the resolution. The look of the
      image should be just the same as if you opened it in Inkscape, for
      example. The issue mentioned above shouldn't happen.

    b.  if the width and height are defined in points "pt" or any similar
      unit, canvas size will be calculated using the conversion:
      (1 point = 1/72 of inch)
      canvas_size_in_pixels = points_size * dpi / 72.

2. If the SVG is imported into an existing image,
   a.  if the viewbox is defined in pixels "px", then
     it takes the resolution of the image into account when calculating
     the viewbox (but also asks the user for its own resolution, too).
     For example if you have an image with image_dpi = 72,
     and you import svg_dpi = 72, then the vector layer will
     have the exact same bounding rect as if it was imported
     into a new document.
     However if the image_dpi = 72 and svg_dpi = 144, then
     the vector layer dimensions will be twice smaller.

   b.  if the viewbox is defined in points "px", then
     it doesn't take the dpi defined by the user into account.
     I don't see any effect of that setting.
     It does seem to take the image dpi into account.