Android: Make Krita a singleTask application
With this if a user opens a file supported by Krita
from an external app, rather than launching it in a
separate instance of Krita (which will freeze). This
will launch it in same instance and then we pass it to
fileOpenRequested.

This is similar to a browser app, which opens "new tab"
rather than new activites.

(cherry picked from commit 44e0b1677ecfa398305d9df5090f2d90daf67fe6)