Draw reference images in high res on 4k display
Before this commit, reference images would be stored in a buffer
which size was not determined by the actual size of the area in pixels
which was available for the QPainter, but that size scaled down by
devicePixelRatioF(). Hence on high resolution displays all
reference images would look blurry and low-res.

This commit fixes this behaviour by introducing high-res buffer
and making sure that it takes the reference image from the mipmap
which size is the closest to the actual new size instead of the
one that is scaled down because of the UI scaling
(devicePixelRatioF()).

I also changed the way the pixel art images (or any images that are being enlarged)
are drawn: now they use FastTransformation instead of SmoothTransformation.

BUG:411118
(cherry picked from commit 1d8b94e3c545b5d2ac98242dd14c6c0becb17071)