Remove the workaround border from the ref images
Before this commit, Reference Images would use the same workaround
as the brush tips since they all use KisQImagePyramid.
However it is not needed for Reference Images and creates
an empty margin around the images.

This commit adds a new method to KisQImagePyramid to get the image
without the workaround border.
(The workaround is just 1 pixel of transparency around the whole edge
that KisQImagePyramid is returning as the closest mipmap image).

CCBUG:411118