Draw resource tooltips in high res in resource selector
Before this commit, resource tooltips in resource selector would be
drawn in the resolution *after* UI scaling applied.
This commit ensures that they are drawn in high resolution
by making sure that the QImage is aware of the UI scaling
and that the scaling to fixed size won't scale it too much.