Android: Fix input event interruptions on canvas.
The explanation of the commit is given in the patch file.
In short, we would stop receiving events due to any floating
message, color etc this would leave canvas in an undefined
state and when we restart the event sequence, canvas would
jump to an artbitrary point or create random lines.

# Conflicts:
#	3rdparty/ext_qt/CMakeLists.txt