Fix no scaling of brush tips in resource item delegate
Before this commit, small brush tips would not be scaled up
in the resource item delegate (used in Resource View in Database
Explorer). This commit makes sure they are scaled up using the
FastTransformation which allows the user to see what the resolution
is (because it shows pixels; SmoothTransformation looks just as low-res
but doesn't provide the information how much low-res it is).