Fix MyPaint shade selector changing look on resizing
Before this commit, MyPaint shade selector would look differently
depending on the ratio of height and width (horizontal stripes)
or depending on the actual pixel length of the stripe
(diagonal stripes).

This commit fixes that behaviour by making sure that the ratio at which
colors change is consistent with different sizes of MyPaint selector.
The ratio is now calculated using the proportion of the distance
of the pixel from the center, to the distance between the center
to the closest edge.


(cherry picked from commit eb6570622805916d39653cec415fd99e93c94952)