Fix ugly edge of MyPaint shade selector
Before this commit, the MyPaint shade selector would have an ugly edge
on the inside circle. That was because the anti-aliasing is done
manually, and the saturation value was the same for pixel of the inside
circle and the pixel of the background, even though the pixel of the
background should have the saturation value reset to the saturation of
the original color.
This commit fixes this behaviour by resetting the saturation value
to the saturation of the original color (note that h, s, and v in the
code represent changes, not absolute values; hence resetting to 0
is the correct behaviour).


(cherry picked from commit 085cdbaae4aae7b2aa10b13f58b6b9d9b2e744e1)