Draw Popup Palette in high res on high dpi displays
Before this commit, both the brush icons and the triangle color selector
inside the popup palette was drawn in low resolution on high dpi
displays.
This commit ensures all of it is drawn with the highest possible
resolution.

This commit also contains a change in a QRegion defining a mask for the
color selector inside. This is necessary for the anti-aliasing to work
nicely. If the Ellipse region is too small, its rough low-res
anti-aliasing (if there is any anti-aliasing at all) makes the
anti-aliasing on high dpi displays look really bad. This commit makes
the ellipse significantly bigger and intersects it with a QRegion
to make sure the mask doesn't go outside of the QWidget area.

(cherry picked from commit 1467dc3520472e03cf0a4963fc1d053fb33d1986)