Draw high res thumbnails in Last Documents docker
Before this commit, Last Documents docker Python plugin would contain
low resolution thumbnails. This commit ensures that the thumbnails are
drawn with the highest possible resolution on high dpi displays.

This commit also contains a fix for pixel art images to be drawn using
the Qt.FastTransformation algorithm instead of SmoothTransformation.
Note that for it to work with .kra files, it needs to be a
'mergedimage.png' scaled up, not 'preview.png', because 'preview.png'
is already smoothed and scaled.