Fix malfunction of x/y position spin boxes in move tool's options
The signal connections had to be moved from the constructor to
activate(), since they get disconnected on every deactivate().

Also, manipulating the x or y position without clicking on the canvas first
doubled the initial value, because a new stroke gets created, which in turn
resets the handle rect.
Reading the value before starting the stroke and making sure it's not
currently unavailable until recalculated should prevent it.

BUG:420329,423452