Draw Comic Manager pages in high res on high dpi displays
Before this commit, images shown by Comic Manager would be drawn
with low resolution on high dpi displays. This commit ensures
that they use the highest resolution available.

Note: because the pages in Comic Manager Docker
(as opposed to the Page Viewer) use "preview.png"
(Page Viewer uses "mergedimage.png"), which size is
limited to max 256 pixels in one dimension,
if the thumbnails size is big or if the file is pixel art,
they won't look good anyway, because then the thumbnails size
would be bigger than 256x256, so the thumbnails would be the result
of upscaling from 256x256 to the thumbnalSize size.

However the change does make it look better in the Page Viewer
(including Pixel Art) and in the main docker for smaller thumbnails
(unfortunately not for Pixel Art because the preview.png is scaled
using the Smooth Transformation already).