Workaround a deadlock when painting on a vector selection with Instant Preview
Instant Preview is based on running two consequent with different
level of detail, but our vector selection don't support that. More
than that, when starting a paint operation on a vector selection,
it should be flattened into a pixel selection. Which means that in LoD
mode the flattening will happen twice.

Ideally, we should somehow make KisSelection survive flattening in
two separate LoD strokes, but I'm not sure it is really worth the
effort.

The downsides of the patch are:

1) The first painting stroke on a selection based layer or
   mask with vector selection will always be executed without
   instant preview. It affects all mask, generator layers and
   filter layers.

2) The first applicaiton of a filter on such a layer will also
   be slow.

BUG:428260
CC:kimageshop@kde.org