Fix loading masks with vector selections
Loading has become broken after commit from June 2020:
6f283ca981 Fix undo breakage after converting a pixel selecion into a vector one

After this change, when a selection is converted into a vector form,
the content of the pixel selection becomes discarded (due to a separate
undo history). Theoreticalll, the projection should become updated using
KisUpdateSelectionJob, but it happens too lately (because the image is
locked while loading). More than that, this job doesn't cause a setDirty()
for the parental node. It happens because we don't usually expect vector
selections to be edited on anything other than selection masks (it is just
not possible for anything other than selection masks).

This patch fixes the issue by calling selection->updateProjection()
explicitly. That is not ideal, because it makes the projection
recalculated twice, but it doesn't seem to be too critical.

The patch also skips loading pixel selection at all, if any vector
selection is present. There is no need for it anymore.

BUG:428332