Generate pixelart previews using FastTransformation
Before this commit, all previews for .kra files were generated
using the Bicubic algorithm.
This commit checks if there is a chance that the file is a pixel art
file, and in such case creates a preview using the FastTransformation
(presumably Nearest Neighbour, but it isn't said in the Qt
documentation) algorithm.

It is determined that it is safe to use FastTransformation algorithm
in case when both of those conditions are true:
- the file is smaller than the preview size
- the exact scale in both dimensions is integer (so it will enlarge
all pixels the same way).

(cherry picked from commit ab3890d0bc70587ab5e5e7a75c11d2d4e5e3ded2)