Add OS-specific action property retrieval to KisActionRegistry
This enables the retrieval of operating system-specific action
properties.

OS-specific action properties are labeled with the attributes
"operatingSystem" (for the selected OS; currently only "macos" is
defined) and "operatingSystemElse" (for the rest).

This commit adds a function, getChildContentForOS, that:
- When finding a matching "operatingSystem", returns the given node's
text
- When finding a node with "operatingSystemElse", stores the last seen
one
- When finding a node without any of these parameters, stores the first
seen one and ignores the rest

This commit implements said change on the Del and Backspace key
shortcuts for macOS. See krita.action.

BUG: 425370
CCMAIL: kimageshop@kde.org