Android: Fix .kra/.ora corruption on re-saves
The problem: WriteOnly flag on QPA Android wasn't being used correctly
as a result of this -- while editing a file if the first save is let's
say 1MB and the second save is supposed to be 512KB, when overwriting
the second file only the first 512KBs get overwritten and the remaining
file is intact, but QuaZip doesn't necessarily knows this. Which creates
a corrupt zip. Once again: previous artifact isn't truncated in
WriteOnly mode when it *should* be.


(cherry picked from commit 034d9a7e8e88db206f572b7c3154737dec1882c2)