Change to the way storyboards and keyframes interact
The original design caused the user to fight with the storyboard docker when
trying to animate. This new design should help resolve conflicts between
animation workflow and storyboarding workflow. Additionally, some QOL
features have been added or worked on such as refinements to scene
reorganization.

The lock button is currently not used, but I would like to use that to
freeze all keyframes to their current position regardless of changes made
to scene duration / position.