Fix assert when requesting liquify mode on an empty layer
The assert was rather harmless, it just checked if the liquify
mode was initialized or not. Before the recent changes in the
transform tool, currentArgs.isIdentity() could not be called
in case of sckipped initialization, but now it is possible. Just
consider the transformation as identity in case initialization was
skipped.

BUG:428685

# Conflicts:
#	plugins/tools/tool_transform2/tool_transform_args.cc