Fix triangle color selector not reacting to mouse events
Before this commit, there was an issue with selecting the colors from
the left and top areas. It was possible to move to them only if the user
dragged the point from the right-bottom area. This was a regression
from ae1b3c7189.

This commit fixes this behaviour by making sure that the "constainsPoint"
method returns true for all the appropriate points even on high dpi
displays.