Upgrade macOS integration plugins
This commit brings a massive refactor to the existing plugins for macOS:

- Adds QuickLook Thumbnailing (10.15+) support.
- Adds Spotlight metadata support.
- It drops NSTasks for reading the KRA file in favor of a bundled
minizip library. NSTask is usable from QuickLook generators, but not
from Spotlight (mdimporter/mdworker).
- It adds license headers throughout the updated code, recognising the
original authors of the Quicklook plugin.

The README and packaging scripts have been updated to reflect these changes.
The LICENSE (GPLv2) has been removed to make it consistent with the rest
of Krita; feel free to contact me if this is wrong.

CCMAIL: kimageshop@kde.org