Correctly check currently active text editor
Previously, we would only check the index of the tab we're on,
which would work as expected in `Rich Text Only` and `Both` editor modes.
But in `SVG Source Only` mode, it would mistakenly think we're on
rich text editor, instead of SVG editor.

That breaks saving, as we were trying to save text document instead of SVG.
Would crash in `KoSvgTextShapeMarkupConverter::convertDocumentToSvg`
as the text document would have no text.
Or empty the text object if we try to avoid the crash.