Imlement first version of Instant preview for the transform tool
It is not an "Instant Preview" in classical terms. That is, it doesn't
run two parallel strokes with the same data. Instead, a portion of
KisSyncLodCacheStrokeStrategy has been integrated into Transform Stroke
Strategy. That is, the transform stroke tries to run LodN jobs while
being a "legacy" stroke itself.

To achieve that, KisStrokesQueue's logic has been modified. Previously,
each stroke had a fixed levelOfDetail it worked on. Now the granularity
is increased, and level of detail can be assigned to each individual
stroke job. By default, job's levelOfDetail is inherited from the stroke's
one, so all the jobs have the save levelOfDetail. But if the stroke has
enough self-confidence, then it may creates jobs working on another
level of detail. In such a case, the stroke should prepare the lod
cached manually by executing jobs from
KisSyncLodCacheStrokeStrategy::createJobsData