Add imageRect to defaultPixel selections in Similar Col Sel Tool
Before this commit, the similar color selection tool would calculate
the selection based only on the paint device content, ignoring the image
bounds. This commit ensures that imageRect is always included for
calculations.

Note that the imageRect still won't be selected if the pixels are of
different color. It only affects the case when for example the user
clicks on a transparent pixel and the non-transparent pixels on that
layer are only located on the top of the layer; assuming that
defaultPixel color is black-transparent, as usual for new layers,
the selection search area will be extended to cover the whole image
to allow for selection to cover the lower parts where all pixeols are
transparnet, but are outside of the paint device bounds.