Fix artifacts on Similar Col Sel Tool on transparent layers
Before this commit, in some cases selection would be very uneven
and unexpected by the user.

Test case:
- transparent layer
- brush: Basic-5 Size
- color: green (anything very different from black will do)
- paint a dot
- Fuzziness: 1
- mode: Current Layer

That was caused by the fact that some pixels would be painted
with green and have opacity = 0. In case of Fuziness = 1,
the whole color is checked using memcmp method.
For transparent pixels it doesn't matter which color they are painted
with, it's enough that they are all transparent (for the user and for
Krita all fully transparent pixels are treated the same way, as having
one and the same color).
Hence this commit adds a check for just opacity in case of transparent
pixel.