Remember last used mode in Sim Col Sel Tool
Before this commit, every time a user would open a new image,
the Similar Color Selection Tool would reset to the Current Layer
option. This commit remembers user's decision regarding the mode.
Note: the color labels are not remembered, because they depend
on the image. It could probably be remembered and restored in some
smart way, but it's not done yet.