Fix isometric grid drawing without any offset
Before this commit, the isometric grid would start from both
the top left and top right corners, but those lines would actually go
upwards, so for the user, it would just look like random offsets.
This commit moves the lines topleft/bottomright with such offset
that it would start exactly in the (0,0) top-left corner, which makes
the grid much more understandable for users and it lets it be always the
same when user extends or shrinks the canvas starting from top-left
corner (which is consistent with the rectangular grid).

This commit changes the offset of isometric grid with the same
settings, so it might break user files. However before recent changes,
the isometric grid would be always incorrect in case the user
zooms in the image (when the updateArea doesn't fit the whole canvas)
so it's better to do those changes now than later.

There is no snapping with isometric grid. The offsets moves the grid
exactly as they did before my changes.