Fix isometric grid with offsets
Before this commit, one offset would move the crosses between lines
upwards (even though the offset was in positive direction),
and the other offset would move it upwards-right (diagonally).
This commit moves them this way: if it's X offset, it moves the crosses
to the right (just like the rectangular grid does), and if it's Y
offset, it moves the crosses downwards (like the rectangular grid).

It makes it more user-friendly since it brings consistency with the
rectangular grid and more sense to the X/Y descriptions of the offset.