Avoid excessive copying of paint devices when applying transformations
All transform workers used to create internal copis of paint devices,
which was excessive, because the calling side has did already have some
copies prepared (for a different reason). Now those internal copies are
gone. The only drawback of this move is that now it might be not that
clear that source and destination paint devices should have exactly the
same color space. Therefore, there are asserts now.