Python Plugin Importer can now import directly from Web
Add a new action to the Tools -> Scripts menu: "Import Python Plugin
from Web"

This lets you paste a URL from which to download and import the
plugin. The URL can either be a link to a Github repository, in which
case the plugin will automaticallly determine the zip download
location, or it can be a direct link to a zip file on any website.

The existing plugin import action has been renamed to "Import Python
Plugin from File"