[FEATURE] Use Shift modifier to avoid "continued transform mode"
When starting a transform action on the same layer multiple times
Krita, by default, tries to restore the state of the previous
transformation. To avoid that, the user had to prss Esc key once,
then teh transform would be restarted from the beginning.

This patch introduces a new feature: after finishing the first
transformation, one can Shift+click on the layer and the transformation
will be started anew, without trying to continue the prevvious one.