Change removeRows, add removeItem and reimplement undo
removeItem will be used to remove scenes. All keyframes
within the scene will be deleted and the keyframes belonging
to scenes after will be moved to fill in the gap. This change
will be undoable. Undo for Add scene will remove the keyframe
for the scene and move keyframes after to fill in the gap.