Fix issues with blending in the mypaint op, break mirror mode for it.
This patch replaces the manual sampling with just using the colormixop
that we have already. It removes my attempt at making compositeops work
for this engine, as that just isn't compatible with mypaint's mixing.

It does implements it's own eraser mode. It also makes sure that color
mixing works in the floating point spaces.

Because I cannot figure out how to make a fixed paintdevice for the mypaint
mask, I had to remove mirror mode.

Fixes the test, also fixes the weird whitening, also fixes the fact that
the paintop was sampling from the image projection instead of the active layer.

BUG:430380