Disable QuickLook Thumbnailing extension
It is the only extension that applies to Catalina onwards, and the bug
report says it takes down the Finder. As I don't have a crash report to
fix, let's just disable it for now.

It'll have to be revisited in the future, because QuickLook generators
themselves are deprecated.

BUG: 430553
(cherry picked from commit 4165f6fbf84a32364fd0d155c39baca8fbf07412)