Android: fix app freeze problem caused by pop-up widgets
Reason: If we double press KisPopupButton, for our first press we tell
Qt to widget->show() which casues an internal flush of event queue then
the second event is sent down which hides the widget.

The second event is immature because the actions of first event didn't
complete in time. This makes the widget hidden while also grabbing mouse
and keyboard, thus freezing the entire app.

Note: The bug is a very time sensitive.

BUG:429636