Fix initiating a pattern name when it's not in the file
Before this commit, patterns without a name in the .pat file
would be initialized in the database with empty names
despite the fact that other resources use their filenames
as the default name (after excluding the extension).
Then there would be an attempt to translate those names,
which would result in:
- multiple warnings:
   "Trying to convert empty KLocalizedString to QString."
- database entries for tooltip saying:
   "(I18N_EMPTY_MESSAGE)"

This commit fixes that by ensuring that the name of the resource
is not replaced by an empty name that is read from the .pat file.