Fix engine sorting in brush presets popup
Before this commit, brush presets popup sorting by Engine
wouldn't work, because it was based on metadata of resources,
and no resources had any metadata (only bundle metadata was saved).
This commit ensures that the metadata of the preset are saved into
the database and that the 'paintopid' metadata is loaded from file
and saved into KisPaintOpPresetSP (so then it can be saved into the
database).

Note: to see it working, you need to reset your database since
it puts the data there on adding the resources.