Fix exporting asl from the current layer style
Before this commit, exporting an asl file from the current layer style
was not implemented. This commit implements it using the
KisAslLayerStyleSerializer class.

Note that Krita cannot import layer styles yet, only export. But the
exported file is imported correctly in Krita 4.4.2-beta2.

This commit also updates the list of storages in the combobox.