Load layer styles on open kra to the memory storage
Before this commit, layer styles from kra file wouldn't load at all.
This commit ensures that they do load to the memory storage.
They will disappear from the model when the user closes the document.

Note: they are not assigned to layers yet; but if you open a file with
a layer style, you can assign the layer style yourself and you'll see
it's the same one.