Fix layer styles md5 sum calculating
Before this commit, md5 sum of a layer style couldn't be calculated
properly because it was based only on an empty byte array.
This commit introduces proper md5 using an ASL serializer
that serializes a collection of layer styles which includes
only the layer style in question. A md5 defined this way should be
the same for the same layer style and usually different for different
ones.